Split code into separate file
